Keep it clean!

+++ 100-clean_web_static.py
@@ -5,6 +5,7 @@ from fabric.api import *
 
 env.hosts = ['52.201.190.114', '52.86.172.121']
 
+
 def do_clean(number=0):
     """Delete out-of-date archives.
     Args:
@@ -16,12 +17,14 @@ def do_clean(number=0):
     number = 1 if int(number) == 0 else int(number)
 
     archives = sorted(os.listdir("versions"))
-    [archives.pop() for i in range(number)]
-    with lcd("versions"):
-        [local("rm ./{}".format(a)) for a in archives]
+    if len(archives) >= number:
+        [archives.pop() for i in range(number)]
+        with lcd("versions"):
+            [local("rm ./{}".format(a)) for a in archives]
 
     with cd("/data/web_static/releases"):
         archives = run("ls -tr").split()
         archives = [a for a in archives if "web_static_" in a]
-        [archives.pop() for i in range(number)]
-        [run("rm -rf ./{}".format(a)) for a in archives]
+        if len(archives) >= number:
+            [archives.pop() for i in range(number)]
+            [run("rm -rf ./{}".format(a)) for a in archives]
